The 1:34AM Update
Movement was handled, should be snappier and handle gravity better.

A new bounce audio that isn't working just yet has been added. I tried using an AudioManager script to do this but I might just put a spot inside the playercontroller script and make the noise there.

The ButtonManager script is useless and doesn't work at the moment, I'll probably delete it.

I touched up on the code for the grow and shrink pads, now allowing them to change the player to normal-size when hitting the opposite pad if you're already big or small.